Move Edit NFTs section of nft draw to own comp
