+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/ReconfigureFundingCycleSettingsPage/ReconfigurePreview/NftSummarySection.tsx
@@ -2,7 +2,7 @@ import { Trans } from '@lingui/macro'
 import { Col, Image, Row } from 'antd'
 import ExternalLink from 'components/ExternalLink'
 import Paragraph from 'components/Paragraph'
-import { NFT_IMAGE_SIDE_LENGTH } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer/NftUpload'
+import { NFT_IMAGE_SIDE_LENGTH } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer/EditNftsSection/NftRewardTierModal/NftUpload'
 import { DEFAULT_NFT_MAX_SUPPLY } from 'contexts/NftRewards/NftRewards'
 import { V2V3ProjectContext } from 'contexts/v2v3/Project/V2V3ProjectContext'
 import { useContext } from 'react'
